Fix implicit type conversion warnings
- Add explicit static_cast<int> for float-to-int conversions in pngyu_util.h
- Add explicit static_cast for size calculations in executecompressthread.cpp and executecompressworkerthread.cpp
- Add explicit static_cast for container size conversions in pngyumainwindow.cpp
- Add const correctness to execute_compress_all parameter in pngyumainwindow.h
- Add strict compiler warning flags (-Wall, -Wextra, -Wconversion, etc.) to CMakeLists.txt

+++ src/executecompressthread.cpp
@@ -40,7 +40,7 @@ QByteArray ExecuteCompressThread::output_png_data() const {
 }
 
 int ExecuteCompressThread::saved_size() const {
-  return m_src_png_data.size() - m_dst_png_data.size();
+  return static_cast<int>(m_src_png_data.size() - m_dst_png_data.size());
 }
 
 QString ExecuteCompressThread::error_string() const { return m_error_string; }
